Move uppercase icon into it's own directory
