+++ src/formats/uppercase/icons.js
